Added additional code to avoid false interrupt counts.
The problem was that the interrupt generator (pulse counter) and the controller were connected through a long cable + low pass filter at the side of the counter. 

This could cause problems through transients that were induced on the cable -> in some cases this caused false interrupts counts.

Now, the counter gets incremented and the interrupt is disabled for ~500 ms.