Receiving of messages is difficult. The function high_isr which is handling the serial interrupt is taking to long and we loose bytes.
For this some code was uncommented and in the function xpl_fifo_push_byte the if statement was uncommented. Due to this no flow control is done but the fifo buffer is set to 100 bytes which is enough to handle the whole xpl message.

Solution is changing the UART speed from 38400 to 19200 but I do not know how to do this. Lieven you will need to help me out...